Begin code cleanup for release
