+++ app/src/main/java/ultrapoulet/wifeygame/battle/BattleEnemy.java
@@ -184,7 +184,7 @@ public class BattleEnemy implements BattleCharacter{
         if(powerDownActive){
             multiplier -= powerDownPercentage;
         }
-        System.out.println("Enemy's increasing damage by: " + multiplier);
+        System.out.println("Enemy's multiplying damage by: " + multiplier);
         int modDamage = (int) (baseDamage * multiplier);
         modDamage = modDamage + (int) ((modDamage / 10) * Math.random());
         return modDamage;
@@ -200,7 +200,7 @@ public class BattleEnemy implements BattleCharacter{
         if(powerDownActive){
             multiplier -= powerDownPercentage;
         }
-        System.out.println("Enemy's increasing damage by: " + multiplier);
+        System.out.println("Enemy's multiplying damage by: " + multiplier);
         int modDamage = (int) (baseDamage * multiplier);
         modDamage = modDamage + (int) ((modDamage / 10) * Math.random());
         return modDamage;
@@ -216,7 +216,7 @@ public class BattleEnemy implements BattleCharacter{
         if(powerDownActive){
             multiplier -= powerDownPercentage;
         }
-        System.out.println("Enemy's increasing damage by: " + multiplier);
+        System.out.println("Enemy's multiplying damage by: " + multiplier);
         int modDamage = (int) (baseDamage * multiplier);
         modDamage = modDamage + (int) ((modDamage / 10) * Math.random());
         return modDamage;
@@ -240,7 +240,7 @@ public class BattleEnemy implements BattleCharacter{
         if(powerDownActive){
             multiplier -= powerDownPercentage;
         }
-        System.out.println("Enemy's increasing damage by: " + multiplier);
+        System.out.println("Enemy's multiplying damage by: " + multiplier);
         int modDamage = (int) (baseDamage * multiplier);
         modDamage = modDamage + (int) ((modDamage / 10) * Math.random());
         return modDamage;
