--- app/src/main/java/ultrapoulet/wifeygame/battle/skills/SkillFactory.java
@@ -1,11 +1,0 @@
-package ultrapoulet.wifeygame.battle.skills;
-
-/**
- * Created by John on 7/12/2016.
- */
-public class SkillFactory {
-
-    public static AbsSkill getSkill(String skillname){
-        return null;
-    }
-}
