+++ app/src/main/java/ultrapoulet/wifeygame/battle/skills/implementations/FujoshiSkill.java
@@ -1,25 +1,34 @@
-package ultrapoulet.wifeygame.battle.skills;
+package ultrapoulet.wifeygame.battle.skills.implementations;
 
 import ultrapoulet.wifeygame.battle.BattleCharacter;
+import ultrapoulet.wifeygame.battle.skills.AbsSkill;
 
 /**
- * Created by John on 8/10/2016.
+ * Created by John on 8/13/2016.
  */
-public class DetectiveSkill extends AbsSkill {
+public class FujoshiSkill extends AbsSkill {
 
-    private double startMultiplier = 0.5;
-    private double perTurn = 0.25;
-    private double maxMultiplier = 5.0;
-    private double multiplier;
-
-    public DetectiveSkill(BattleCharacter owner) {
+    public FujoshiSkill(BattleCharacter owner){
         super(owner);
-        this.skillName = "Detective";
+        this.skillName = "Fujoshi";
     }
 
+    private double multiplier = 1.0;
+    private double perTrap = 1.0/3.0;
+    private double twoTrapBonus = 1.0;
+
     @Override
-    public void startWave() {
-        multiplier = startMultiplier;
+    public void startBattle(BattleCharacter[] party) {
+        int count = 0;
+        for(int i = 0; i < party.length; i++){
+            if(party[i] != owner && party[i].hasSkill(TrapSkill.class)){
+                count++;
+            }
+        }
+        multiplier += count * perTrap;
+        if(count >= 2){
+            multiplier += twoTrapBonus;
+        }
     }
 
     @Override
@@ -37,12 +46,6 @@ public class DetectiveSkill extends AbsSkill {
         return multiplier;
     }
 
-    @Override
-    public void endRound() {
-        if(multiplier < maxMultiplier) {
-            multiplier += perTurn;
-        }
-    }
 
     @Override
     public double[] getMultipliers(BattleCharacter enemy) {
@@ -60,8 +63,8 @@ public class DetectiveSkill extends AbsSkill {
     @Override
     public String getDescription(BattleCharacter enemy) {
         StringBuilder desc = new StringBuilder();
-        desc.append("Attack Multiplier: " + multiplier + "\n\n");
-        desc.append("Multiplies damage dealt by 0.5x at the start of a wave. Multiplier increases by 0.25x at the end of each turn, up to a maximum of 5.0x.");
+        desc.append("Attack Multiplier: " + String.format("%1$.2f", multiplier) + "\n\n");
+        desc.append("Increases damage dealt multiplier by 0.33x for each Trap wifey in the party. Damage increased by a bonus 1.0x if there are more than two Trap wifeys.");
         return desc.toString();
     }
 }
