+++ app/src/main/java/ultrapoulet/wifeygame/battle/skills/implementations/MaidSkill.java
@@ -1,6 +1,7 @@
-package ultrapoulet.wifeygame.battle.skills;
+package ultrapoulet.wifeygame.battle.skills.implementations;
 
 import ultrapoulet.wifeygame.battle.BattleCharacter;
+import ultrapoulet.wifeygame.battle.skills.AbsSkill;
 
 /**
  * Created by John on 7/31/2016.
