+++ app/src/main/java/ultrapoulet/wifeygame/battle/skills/implementations/MediumSkill.java
@@ -1,6 +1,8 @@
-package ultrapoulet.wifeygame.battle.skills;
+package ultrapoulet.wifeygame.battle.skills.implementations;
 
 import ultrapoulet.wifeygame.battle.BattleCharacter;
+import ultrapoulet.wifeygame.battle.skills.AbsSkill;
+import ultrapoulet.wifeygame.battle.skills.implementations.GhostSkill;
 
 /**
  * Created by John on 7/21/2016.
