+++ app/src/main/java/ultrapoulet/wifeygame/battle/skills/implementations/TimeTravelerSkill.java
@@ -1,6 +1,7 @@
-package ultrapoulet.wifeygame.battle.skills;
+package ultrapoulet.wifeygame.battle.skills.implementations;
 
 import ultrapoulet.wifeygame.battle.BattleCharacter;
+import ultrapoulet.wifeygame.battle.skills.AbsSkill;
 
 /**
  * Created by John on 8/2/2016.
