+++ app/src/main/java/ultrapoulet/wifeygame/character/SkillsEnum.java
@@ -4,36 +4,36 @@ import java.util.HashMap;
 
 import ultrapoulet.wifeygame.battle.BattleCharacter;
 import ultrapoulet.wifeygame.battle.skills.AbsSkill;
-import ultrapoulet.wifeygame.battle.skills.AthleteSkill;
-import ultrapoulet.wifeygame.battle.skills.BikerSkill;
-import ultrapoulet.wifeygame.battle.skills.ChuuniSkill;
-import ultrapoulet.wifeygame.battle.skills.DetectiveSkill;
-import ultrapoulet.wifeygame.battle.skills.DullahanSkill;
-import ultrapoulet.wifeygame.battle.skills.FujoshiSkill;
-import ultrapoulet.wifeygame.battle.skills.GhostSkill;
-import ultrapoulet.wifeygame.battle.skills.HyperSkill;
-import ultrapoulet.wifeygame.battle.skills.KillerSkill;
-import ultrapoulet.wifeygame.battle.skills.MagicalGirlSkill;
-import ultrapoulet.wifeygame.battle.skills.MaidSkill;
-import ultrapoulet.wifeygame.battle.skills.MasochistSkill;
-import ultrapoulet.wifeygame.battle.skills.MechanicSkill;
-import ultrapoulet.wifeygame.battle.skills.MediumSkill;
-import ultrapoulet.wifeygame.battle.skills.MusicianSkill;
-import ultrapoulet.wifeygame.battle.skills.NurseSkill;
-import ultrapoulet.wifeygame.battle.skills.PravdaSkill;
-import ultrapoulet.wifeygame.battle.skills.ProgrammerSkill;
-import ultrapoulet.wifeygame.battle.skills.ProtagonistSkill;
-import ultrapoulet.wifeygame.battle.skills.RacerSkill;
-import ultrapoulet.wifeygame.battle.skills.RobotSkill;
-import ultrapoulet.wifeygame.battle.skills.SadistSkill;
-import ultrapoulet.wifeygame.battle.skills.SlugabedSkill;
-import ultrapoulet.wifeygame.battle.skills.SportsManagerSkill;
-import ultrapoulet.wifeygame.battle.skills.SurvivalSkill;
-import ultrapoulet.wifeygame.battle.skills.TimeTravelerSkill;
-import ultrapoulet.wifeygame.battle.skills.TrapSkill;
-import ultrapoulet.wifeygame.battle.skills.TsundereSkill;
-import ultrapoulet.wifeygame.battle.skills.VampireSkill;
-import ultrapoulet.wifeygame.battle.skills.WitchSkill;
+import ultrapoulet.wifeygame.battle.skills.implementations.AthleteSkill;
+import ultrapoulet.wifeygame.battle.skills.implementations.BikerSkill;
+import ultrapoulet.wifeygame.battle.skills.implementations.ChuuniSkill;
+import ultrapoulet.wifeygame.battle.skills.implementations.DetectiveSkill;
+import ultrapoulet.wifeygame.battle.skills.implementations.DullahanSkill;
+import ultrapoulet.wifeygame.battle.skills.implementations.FujoshiSkill;
+import ultrapoulet.wifeygame.battle.skills.implementations.GhostSkill;
+import ultrapoulet.wifeygame.battle.skills.implementations.HyperSkill;
+import ultrapoulet.wifeygame.battle.skills.implementations.KillerSkill;
+import ultrapoulet.wifeygame.battle.skills.implementations.MagicalGirlSkill;
+import ultrapoulet.wifeygame.battle.skills.implementations.MaidSkill;
+import ultrapoulet.wifeygame.battle.skills.implementations.MasochistSkill;
+import ultrapoulet.wifeygame.battle.skills.implementations.MechanicSkill;
+import ultrapoulet.wifeygame.battle.skills.implementations.MediumSkill;
+import ultrapoulet.wifeygame.battle.skills.implementations.MusicianSkill;
+import ultrapoulet.wifeygame.battle.skills.implementations.NurseSkill;
+import ultrapoulet.wifeygame.battle.skills.implementations.PravdaSkill;
+import ultrapoulet.wifeygame.battle.skills.implementations.ProgrammerSkill;
+import ultrapoulet.wifeygame.battle.skills.implementations.ProtagonistSkill;
+import ultrapoulet.wifeygame.battle.skills.implementations.RacerSkill;
+import ultrapoulet.wifeygame.battle.skills.implementations.RobotSkill;
+import ultrapoulet.wifeygame.battle.skills.implementations.SadistSkill;
+import ultrapoulet.wifeygame.battle.skills.implementations.SlugabedSkill;
+import ultrapoulet.wifeygame.battle.skills.implementations.SportsManagerSkill;
+import ultrapoulet.wifeygame.battle.skills.implementations.SurvivalSkill;
+import ultrapoulet.wifeygame.battle.skills.implementations.TimeTravelerSkill;
+import ultrapoulet.wifeygame.battle.skills.implementations.TrapSkill;
+import ultrapoulet.wifeygame.battle.skills.implementations.TsundereSkill;
+import ultrapoulet.wifeygame.battle.skills.implementations.VampireSkill;
+import ultrapoulet.wifeygame.battle.skills.implementations.WitchSkill;
 
 /**
  * Created by John on 7/14/2016.
